connection, jdbc refactored
